requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-braces==1.15.0
django-debug-toolbar==4.3.0
django-embed-video==1.4.9
django-redisboard==8.4.0
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
pymemcache==4.0.0
redis==5.0.4
requests==2.32.3
setuptools==72.2.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
